✨ Support for unary expressions in IR and synthesis algorithm (#259)
## Description

With this PR preliminary support for unary expressions is added to the
intermediate representation as well as for the base synthesis algorithm
(`syrec_synthesis`). As discussed in #243 no support for unary
expressions is added to the SyReC parser.

Fixes #243 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are related to it.
- [X] I have added appropriate tests and documentation.
- [X] I have made sure that all CI jobs on GitHub pass.
- [X] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: Lukas Burgholzer <burgholzer@me.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>