♻️ Replacement of boost::fusion and boost::spirit with ANTLR4 components for SyReC grammar and parser (#220)
## Description

With this pull-request the current boost::fusion and boost::spirit
definition of the SyReC grammar as well as the SyReC parser
implementation is replaced with an equivalent ANTLR4 implementation.
Additionally, various small syntax/semantic issues of the previous
parser are fixed and a more exhaustive specification of the assumed
semantics of the SyReC language are documented.

Fixes one of the open issues of #27 

This pull-request will introduce the following breaking changes to the
public interface of the library:
- The unscoped enumerations used to model the operation component in the
syrec::BinaryExpression, syrec::ShiftExpression, syrec::AssignStatement
and syrec::UnaryStatement are replaced with a scoped enumeration to
improve code readability and to prevent subtle bugs (e.g. performing
assignments using primitive types (i.e. an int) as is the case in the
synthesis code).
- Refactored syrec::Variable _type_ member to use scoped enumeration
instead of primitive type (unsigned int)
- A syrec::Number is now able to model a constant expression as defined
in the SyReC grammar instead of using the syrec::NumericExpression type
(the previous parser allowed all binary operation to be used in a
syrec::NumericExpression with the latter also being used at positions
where a syrec::Number is expected, leading to syntactically incorrect
SyReC programs being considered as valid)
- Added an external dependency: [ANTLR4
runtime](https://github.com/antlr/antlr4/tree/master)
- Added an external dependency: [{fmt}](https://github.com/fmtlib/fmt)
library (temporary solution until std::format of C++20 is available)
- Synthesis of syrec::ForStatement will now assume that an omitted start
value of the iteration range is equal to 0 and does not include the end
value in the iteration range of the loop, i.e. [0, end) instead of the
previous range [1, end].

Non-breaking changes to the public interface include:
- Additional options in the parser configuration object
syrec::ReadProgramSettings
- Option to try and evaluate a syrec::Number instead of crashing if no
value for a loop variable is found
- Reading a SyReC program from a string
(syrec::Program::readFromString()) required for tests of SyReC parser

Other changes included:
- Fixed issues in some circuits used in tests of synthesis algorithms to
consider updated SyReC semantics
- Fixed potential null pointer error when trying to fetch reference
variable of syrec::VariableAccess using getter member function
- Updated mqt.core dependency in pyproject.toml to commit eb47e15 which
includes required bugfix for class qc::Permutation. Most recent commit
of main branch was not chosen to avoid a refactoring of existing code.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are related to it.
- [X] I have added appropriate tests and documentation.
- [X] I have made sure that all CI jobs on GitHub pass.
- [X] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>