Syrec editor can now load SyReC programs from a file (#289)
## Description

With this PR SyReC programs can now be read from a file that is selected
via the 'Open' action in the `syrec-editor` python application. After a
program was opened, the contents of the code editor widget is updated
(and syntax highlighting performed), the error log and circuit view is
cleared. Closing the file dialog opened via the 'Open' action will leave
all Qt widgets unchanged.

Fixes #288 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are related to it.
- [X] I have added appropriate tests and documentation.
- [X] I have made sure that all CI jobs on GitHub pass.
- [X] The pull request introduces no new warnings and follows the
project's style guidelines.