Checking for repeating qubit indices in operands of expression no longer throws index-out-of-range error (#281)
## Description

Prio to the first refactoring of the synthesis algorithm components for
the transition to MQT Core (see issue #29) the check for repeating qubit
indices in the operands of an expression via
`SyrecSynthesis::checkRepeats()` did contain the following check:

```cpp
    bool SyrecSynthesis::checkRepeats() {
        std::vector checkLhsVec(...);
        std::vector checkRhsVec(...);
        ...

        for (auto const& i: checkLhsVec) {
            for (auto const& j: checkRhsVec) {
                if (i == j) {
                    ...
                    return true;
                }
            }
        }
        return false;
    }
```

And was refactoring in #240 to which refactored introduced the invalid
assumption that the `checkLhsVec` and `checkRhsVec` containers have the
same size which is not the case in an expression of the form `(a ^ (b ^
c))` (`checkLhsVec.size()` = 2, `checkRhsVec.size()` = 1) and will lead
to an index-out-of-range access in said expression.

```cpp
    bool SyrecSynthesis::checkRepeats() {
        std::vector checkLhsVec(...);
        std::vector checkRhsVec(...);
        bool foundRepeat = false;
        ...
        for (std::size_t i = 0; i < checkLhsVec.size() && !foundRepeat; ++i) {
            foundRepeat = checkLhsVec[i] == checkRhsVec[i];
        }
        ...
        return foundRepeat;
    }
```

This PR fixes the index-out-of-range access.

Fixes #280 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are related to it.
- [X] I have added appropriate tests and documentation.
- [ ] I have made sure that all CI jobs on GitHub pass.
- [X] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>