Load simulation test case data from .json with unified format (#274)
## Description

As discussed and first introduced in #220, the simulation tests for
single SyReC operations as well as full circuits used two different
mechanisms to load the test data with the former using hard-coded data
in the test case itself while the latter loads the data from a .json
file.

With this PR all simulation tests load their test data from a .json file
that uses the single unified format:

```json
{
   "TEST_NAME": {
      "inputCircuit": "<STRINGIFIED_SYREC_CIRCUIT>",
      "simulationRuns": [
         { "in": "<STRINGIFIED_INPUT_STATE_DATA_QUBIT_VALUES>", 
          "out": "<STRINGIFIED_EXPECTED_OUTPUT_STATE_DATA_QUBIT_VALUES>" }
       ]
   }
}
```
## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are related to it.
- [X] I have added appropriate tests and documentation.
- [X] I have made sure that all CI jobs on GitHub pass.
- [X] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: Lukas Burgholzer <burgholzer@me.com>