Fix of implementation of SyrecSynthesis::increase operation (#286)
## Description

- With this PR the counter-examples found in #254 can be correctly
synthesized due to bugfixes in the `SyrecSynthesis::increase` operation.
Additionally, further tests of the binary `+` operation as well as the
assignment operations `+=` and `++=` are added that check that a
simulation of these operations returns the correct results (the expected
synthesis costs are not defined).

- Various algorithms that can be used fix the errors of
`SyrecSynthesis::increase` were proposed in #254 but since the pre-PR
implementation already followed one of these algorithms (with the
algorithm of "Quantum Addition Circuits and Unbounded Fan-Out"
([Takahashi et al.](https://arxiv.org/abs/0910.2530v1)) being the
closest match) no complete reimplementation was needed.

- Some simulation tests could not be added in the PR due to errors in
the `LineAwareSynthesis` (e.g., see #280)

- A breaking change in the internal `SyrecSynthesis` interface was added
due to the unification of the `SyrecSynthesis::increase` and
`SyrecSynthesis::increaseWithCarry` into one function.

Fixes #254

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are related to it.
- [X] I have added appropriate tests and documentation.
- [X] I have made sure that all CI jobs on GitHub pass.
- [X] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>