🐛 Refactoring of synthesis of division and modulo operations (#305)
## Description

The current implementation of the synthesis of the division and modulo
operations did not create the correct gate sequence and lead to a
missmatch between the expected and actual output values when simulating
the synthesized circuits. With this PR the implementation of both
operations is replaced with the restoring division algorithm defined in
[Quantum Circuit Designs of Integer Division Optimizing T-count and
T-depth](https://arxiv.org/abs/1809.09732). This caused the following
changes:

- Both operands now need 2*N ancillary qubits to store the values of the
quotient and remainder for each operation.
- The internal SyrecSynthesis interface for the division and modulo
operations had to be updated since both
the quotient and remainder qubits need to be provided for both
operations.
- The transistor cost of both operations increased.

Note that the referenced paper also defines a non-restoring division
algorithm that requires less transistors to synthesize. At the moment we
are not sure whether any of the two division algorithms could be
refactored to only require N ancillary qubits (to store the quotient in
the division/remainder in the modulo operation) to synthesis.

Fixes #302 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are related to it.
- [X] I have added appropriate tests and documentation.
- [X] I have made sure that all CI jobs on GitHub pass.
- [X] The pull request introduces no new warnings and follows the
project's style guidelines.