🐛 Synthesis of shift expressions with shift amount larger or equal to bitwidth of to be shift expression. (#297)
## Description

With this PR the synthesis of a shift expression in which the defined
shift amount is larger or equal to the bitwidth of the to be shifted
expression (N) is "simplified" to the creation of N ancillary qubits
(i.e. equal to an substitution of the shift expression with the constant
0) with the latter storing the result of the shift expression.

Fixes #258 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are related to it.
- [X] I have added appropriate tests and documentation.
- [X] I have made sure that all CI jobs on GitHub pass.
- [X] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>