🐛 Fix of duplicate qubit label generation during synthesis (#327)
## Description

This PR introduces the following changes in both the C++ as well as the
python library.
C++:
- Internal qubit labels for ancillary qubits as well as the qubits of
local variables of SyReC modules are generated using the pattern
"__p<CURR_NUM_QUBITS_IN_QUANTUM_COMP>" with the "__p" currently not
being configurable via the synthesis settings. The combinations of
internal qubit label => user defined qubit label are stored in the
`syrec::AnnotatableQuantumComputation`.
- The inline stack of an ancillary/qubits of the local variables of a
SyReC module are optional (activated via a flag in the
`syrec::Properties` passed to the synthesis algorithms) generated and
stored in the `syrec::AnnotatableQuantumComputation` (queriable via
`syrec::AnnotatableQuantumComputation::getInliningInformationOfQubit(...)`),
to provide debug information about the origin of the qubit to the user.
**This feature needs to be explicitly activated with the qubit label
combination generation of the previous bullet point not being effected
by this feature**.
- Three new data structures `syrec::QubitInlineStack`,
`syrec::QubitInliningStack::QubitInliningStackEntry` and
`syrec::AnnotatableQuantumComputation::InlinedQubitInformation` were
added.
- Two new constants `MAIN_MODULE_IDENTIFIER_CONFIG_KEY` and
`GENERATE_INLINE_DEBUG_INFORMATION_CONFIG_KEY` in
`\include\algorithms\syrec_synthesis.hpp` to improve the activation of
synthesis features in `syrec::Properties` instead of relying on
hard-coded strings in the source files.

Python:
- An update to the python bindings to make the newly introduced data
structures available
- The two new string constants are made available as attributes of the
python module but not documented in the reStructured Text files due to
["warnings"](https://google-grain.readthedocs.io/en/latest/_autosummary/grain.constants.RECORD_KEY.html)
being generated in the documentation because no 'doc string' can be
provided for the attributes. This should be fixable in the future if
`.pyi` files are used to document the python package, as is the case in
the `mqt-core` project.
- Two new PyQt6 widgets to update the synthesis settings and display the
qubit inline information:
<img width="540" height="449" alt="labelledEditorWidgets"
src="https://github.com/user-attachments/assets/17c049c0-0db7-4055-bdd4-618dfd982dd0"
/>

The displayed qubit inline information when the feature is
deactivated/activated is the following:
<img width="1041" height="646" alt="qubitInlineInfoDifferences"
src="https://github.com/user-attachments/assets/8a8bcd74-53de-4144-bf63-854180148b3c"
/>

Inline information about a qubit can be selected in the `syrec-editor`
by:
- Selecting the label of an ancillary/local variable qubit of a module
in the combobox
- Click on the label of an ancillary/local variable qubit of a module in
the circuit view

Fixes #278

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are related to it.
- [X] I have added appropriate tests and documentation.
- [X] I have made sure that all CI jobs on GitHub pass.
- [X] The pull request introduces no new warnings and follows the
project's style guidelines.