🐛 Python editor now displays internal qubit label instead of user declared ones for all qubits (#481)
## Description

Instead of fetching the use-declared qubit label for non-ancillary
qubits we always fetch the internal label instead since user declared
qubit labels are only available for the local variables of a SyReC
module which additionally are not unique since as the name suggests
local variables are only unique to the module in which the are declared.
Thus we display the globally unique internal qubit label instead
(improvements to also include the qubit inline information could be made
with #28)

Fixes #480

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are focused and
relevant to this change.
- [X] I have added appropriate tests that cover the new/changed
functionality.
- [X] I have updated the documentation to reflect these changes.
- [X] The changes follow the project's style guidelines and introduce no
new warnings.
- [X] The changes are fully tested and pass the CI checks.
- [X] I have reviewed my own code changes.