🐛 SyReC circuits with synthesis errors are no longer treated as successfully synthesized in python SyReC editor (#486)
## Description

- The return value of the `cost_aware_synthesis` and
`line_aware_synthesis` function of the `mqt.syrec` python module is now
used in the python editor.
- By temporarily redirecting the C++ `std::cerr` outputstream to the
python `sys.stderr` stream for the duration of the execution of the
`cost_aware_synthesis` and `line_aware_synthesis` function of the
`mqt.syrec` python module we are able to capture the synthesis errors
reported by the C++ synthesizers and report them in the python editor.
- Since we want to use a standardized error message output format (`--
line <line> col <col>: <error_msg>`) to report parser/synthesis errors
and due to the reported synthesis errors not containing information
about their origin in the SyReC source code, said synthesis errors are
reported as `-- line ? col ?: <synthesis_error>`.

Fixes #485 <!--- Replace `(issue)` with the issue number fixed by this
pull request. If this PR does not fix an issue, please remove this line.
-->

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [X] The pull request only contains commits that are focused and
relevant to this change.
- [X] I have added appropriate tests that cover the new/changed
functionality.
- [X] I have updated the documentation to reflect these changes.
- [X] The changes follow the project's style guidelines and introduce no
new warnings.
- [X] The changes are fully tested and pass the CI checks.
- [X] I have reviewed my own code changes.